chore: refactor zod schema, lifting inputStyle
