+++ apps/frontend/app/admin/_components/TitleForm.tsx
@@ -1,7 +1,7 @@
 import { Input } from '@/components/ui/input'
 import { cn } from '@/lib/utils'
 import { useFormContext } from 'react-hook-form'
-import { inputStyle } from '../problem/utils'
+import { inputStyle } from '../utils'
 import ErrorMessage from './ErrorMessage'
 
 export default function TitleForm({ placeholder }: { placeholder: string }) {
