+++ apps/frontend/app/admin/problem/_components/LimitForm.tsx
@@ -2,7 +2,7 @@ import { Input } from '@/components/ui/input'
 import { cn } from '@/lib/utils'
 import { useFormContext } from 'react-hook-form'
 import ErrorMessage from '../../_components/ErrorMessage'
-import { inputStyle } from '../utils'
+import { inputStyle } from '../../utils'
 
 export default function LimitForm() {
   const {
