+++ apps/frontend/app/admin/problem/utils.ts
@@ -79,6 +79,3 @@ export const createSchema = commonSchema.extend({
     )
     .min(1)
 })
-
-export const inputStyle =
-  'border-gray-200 focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-gray-950'
